server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
tensorflow==2.13.0
pillow==10.0.0
requests==2.31.0
numpy==1.24.3 